Minor tweak

+++ Receiver.cpp
@@ -58,7 +58,7 @@ int main() {
         std::cout << "Введите цифру в соответствии с желаемым действием:\n"
                      "0 - Завершить работу.\n"
                      "1 - Прочитать сообщение.\n";
-        byte action;
+        short action;
         std::cin >> action;
         if (action == 0) terminate = true;
         else if (action == 1) {
